refactor dsv4 attn
