+++ vllm/model_executor/layers/deepseek_v4_attention.py
@@ -4,7 +4,6 @@
 DeepseekV4 MLA Attention Layer
 """
 
-import math
 from collections.abc import Callable
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any, cast
@@ -28,6 +27,11 @@ from vllm.v1.attention.ops.deepseek_v4_ops import (
     fused_indexer_q_rope_quant,
     fused_inv_rope_fp8_quant,
     fused_q_kv_rmsnorm,
+)
+from vllm.v1.attention.ops.rocm_aiter_mla_sparse import (
+    rocm_forward_decode_fallback,
+    rocm_inv_rope_einsum,
+    rocm_sparse_attn_prefill,
 )
 
 if TYPE_CHECKING:
@@ -210,8 +214,6 @@ class DeepseekV4MultiHeadLatentAttentionWrapper(PluggableLayer):
         self.topk_indices_buffer = mla_modules.topk_indices_buffer
 
         self.indexer = mla_modules.indexer
-        # Keep ROCm on the BF16 reference wo_a path util kernel ready.
-        self.use_ref_wo_a_path = current_platform.is_rocm()
 
         # Per-head RMS normalization for Q (no learnable weights)
         self.q_head_norm = RMSNorm(head_dim, eps=self.eps, has_weight=False)
@@ -306,30 +308,17 @@ class DeepseekV4MultiHeadLatentAttentionWrapper(PluggableLayer):
         )
         o = o_padded[:, : self.n_local_heads, :]
 
-        if self.use_ref_wo_a_path:
-            o_ref = _apply_inv_rope_ref(
-                self.rotary_emb, o, positions, self.rope_head_dim
-            ).to(torch.bfloat16)
-            o_ref = o_ref.view(num_tokens, self.n_local_groups, -1)
-
-            hidden_dim = o_ref.shape[-1]
-            if hasattr(self.wo_a, "weight_scale_inv"):
-                wo_a_weight = self.wo_a.weight.view(
-                    self.n_local_groups, self.o_lora_rank, hidden_dim
-                ).to(torch.float32)
-                wo_a_scale = _expand_2d_block_scales(
-                    self.wo_a.weight_scale_inv.view(
-                        self.n_local_groups, -1, self.wo_a.weight_scale_inv.shape[-1]
-                    ),
-                    self.o_lora_rank,
-                    hidden_dim,
-                )
-                wo_a_weight = (wo_a_weight * wo_a_scale).to(torch.bfloat16)
-            else:
-                wo_a_weight = self.wo_a.weight.view(
-                    self.n_local_groups, self.o_lora_rank, hidden_dim
-                ).to(torch.bfloat16)
-            z = torch.einsum("tgd,grd->tgr", o_ref, wo_a_weight)
+        # Keep ROCm on the BF16 reference wo_a path util kernel ready.
+        if current_platform.is_rocm():
+            z = rocm_inv_rope_einsum(
+                self.rotary_emb,
+                o,
+                positions,
+                self.rope_head_dim,
+                self.n_local_groups,
+                self.o_lora_rank,
+                self.wo_a,
+            )
             return self.wo_b(z.flatten(1))
 
         # O projection: inverse RoPE + FP8 quant + einsum + wo_b
@@ -412,21 +401,14 @@ class DeepseekV4MultiHeadLatentAttentionWrapper(PluggableLayer):
             qr_kv, _ = self.fused_wqa_wkv(hidden_states)
             return qr_kv
 
-        if aux_streams is None:
-            qr_kv = fused_wqa_wkv()
-            kv_score = aux_fns[0]() if aux_fns[0] is not None else None
-            indexer_weights = aux_fns[1]() if aux_fns[1] is not None else None
-            indexer_kv_score = aux_fns[2]() if aux_fns[2] is not None else None
-            return qr_kv, kv_score, indexer_kv_score, indexer_weights
-
         qr_kv, (kv_score, indexer_weights, indexer_kv_score) = execute_in_parallel(
             fused_wqa_wkv,
             aux_fns,
             self.ln_events[0],
             self.ln_events[1:4],
-            self.aux_stream_list[:3],
+            aux_streams,
             enable=hidden_states.shape[0]
-            <= envs.VLLM_MULTI_STREAM_GEMM_TOKEN_THRESHOLD and not current_platform.is_rocm(),
+            <= envs.VLLM_MULTI_STREAM_GEMM_TOKEN_THRESHOLD,
         )
 
         return qr_kv, kv_score, indexer_kv_score, indexer_weights
@@ -612,128 +594,7 @@ def deepseek_v4_fp8_einsum(
     equation: str,
     recipe: list[int],
 ) -> None:
-    try:
-        fp8_einsum(equation, (a, a_scale), (b, b_scale), out, recipe=tuple(recipe))
-    except RuntimeError as exc:
-        if "DeepGEMM backend is not available or outdated" not in str(exc):
-            raise
-        _deepseek_v4_fp8_einsum_fallback(a, a_scale, b, b_scale, out, equation)
-
-
-def _decode_e8m0_scales(scale: torch.Tensor) -> torch.Tensor:
-    if scale.dtype == torch.float8_e8m0fnu:
-        from vllm.model_executor.layers.quantization.utils.fp8_utils import (
-            _upcast_e8m0_to_fp32,
-        )
-
-        return _upcast_e8m0_to_fp32(scale).contiguous()
-    return scale.to(torch.float32)
-
-
-def _expand_last_dim_scales(scale: torch.Tensor, last_dim: int) -> torch.Tensor:
-    scale = _decode_e8m0_scales(scale)
-    block = math.ceil(last_dim / scale.shape[-1])
-    return torch.repeat_interleave(scale, block, dim=-1)[..., :last_dim]
-
-
-def _expand_2d_block_scales(
-    scale: torch.Tensor,
-    rows: int,
-    cols: int,
-) -> torch.Tensor:
-    scale = _decode_e8m0_scales(scale)
-    row_blocks, col_blocks = scale.shape[-2:]
-    row_block = math.ceil(rows / row_blocks)
-    col_block = math.ceil(cols / col_blocks)
-    scale = torch.repeat_interleave(scale, row_block, dim=-2)[..., :rows, :]
-    scale = torch.repeat_interleave(scale, col_block, dim=-1)[..., :, :cols]
-    return scale
-
-
-def _apply_gptj_inv_rope_ref(
-    x: torch.Tensor,
-    positions: torch.Tensor,
-    cos_sin_cache: torch.Tensor,
-    rope_dim: int,
-) -> torch.Tensor:
-    if rope_dim == 0 or x.numel() == 0:
-        return x
-    half_rot = rope_dim // 2
-    nope_dim = x.shape[-1] - rope_dim
-    dtype = x.dtype
-    x = x.to(torch.float32)
-    cache = cos_sin_cache.index_select(0, positions.to(torch.long))
-    cos = cache[:, :half_rot].to(torch.float32)
-    sin = cache[:, half_rot : 2 * half_rot].to(torch.float32)
-    view_shape = (positions.shape[0],) + (1,) * (x.dim() - 2) + (half_rot,)
-    cos = cos.view(view_shape)
-    sin = sin.view(view_shape)
-    rope = x[..., nope_dim:]
-    y_even = rope[..., 0::2]
-    y_odd = rope[..., 1::2]
-    rope_out = torch.stack(
-        (y_even * cos + y_odd * sin, y_odd * cos - y_even * sin),
-        dim=-1,
-    ).flatten(-2)
-    x = x.clone()
-    x[..., nope_dim:] = rope_out
-    return x.to(dtype)
-
-
-def _apply_inv_rope_ref(
-    rotary_emb: torch.nn.Module,
-    x: torch.Tensor,
-    positions: torch.Tensor,
-    rope_dim: int,
-) -> torch.Tensor:
-    if hasattr(rotary_emb, "forward_native"):
-        try:
-            query, _ = rotary_emb.forward_native(
-                positions,
-                x.clone(),
-                None,
-                inverse=True,
-            )
-            return query
-        except TypeError:
-            pass
-    return _apply_gptj_inv_rope_ref(x, positions, rotary_emb.cos_sin_cache, rope_dim)
-
-
-def _deepseek_v4_fp8_einsum_fallback(
-    a: torch.Tensor,
-    a_scale: torch.Tensor,
-    b: torch.Tensor,
-    b_scale: torch.Tensor,
-    out: torch.Tensor,
-    equation: str,
-) -> None:
-    if equation != "bhr,hdr->bhd":
-        raise RuntimeError(f"Unsupported fallback equation: {equation}")
-
-    num_groups = a.shape[1]
-    hidden_dim = a.shape[2]
-    output_dim = b.shape[0] // num_groups
-
-    if b.shape[0] % num_groups != 0:
-        raise RuntimeError(
-            f"Cannot reshape weight of shape {tuple(b.shape)} into "
-            f"({num_groups}, {output_dim}, {hidden_dim})."
-        )
-
-    a_deq = (a.to(torch.float32) * _expand_last_dim_scales(a_scale, hidden_dim)).to(
-        torch.bfloat16
-    )
-
-    b_deq = b.view(num_groups, output_dim, hidden_dim).to(torch.float32)
-    b_scale_deq = _expand_2d_block_scales(
-        b_scale.view(num_groups, -1, b_scale.shape[-1]),
-        output_dim,
-        hidden_dim,
-    )
-    b_deq = (b_deq * b_scale_deq).to(torch.bfloat16)
-
-    out.copy_(torch.einsum(equation, a_deq, b_deq).to(out.dtype))
+    fp8_einsum(equation, (a, a_scale), (b, b_scale), out, recipe=tuple(recipe))
 
 
 def deepseek_v4_fp8_einsum_fake(
@@ -830,11 +691,8 @@ class DeepseekV4MLAAttention(nn.Module, AttentionLayerBase):
             vllm_config.scheduler_config.max_num_batched_tokens
         )
         self.max_model_len = vllm_config.model_config.max_model_len
-        # DeepseekV4 only supports fp8 kv-cache format for now. Treat "auto"
-        # as the model default and normalize it before the fp8-only checks.
+        # DeepseekV4 only supports fp8 kv-cache format for now.
         kv_cache_dtype = cache_config.cache_dtype if cache_config is not None else "fp8"
-        if kv_cache_dtype == "auto":
-            kv_cache_dtype = "fp8"
 
         assert kv_cache_dtype.startswith("fp8"), (
             f"DeepseekV4 only supports fp8 kv-cache format for now, "
@@ -982,15 +840,20 @@ class DeepseekV4MLAAttention(nn.Module, AttentionLayerBase):
         swa_lens = swa_metadata.decode_swa_lens
 
         if current_platform.is_rocm():
-            self._forward_decode_fallback(
+            rocm_forward_decode_fallback(
                 q=q,
                 kv_cache=kv_cache,
-                swa_metadata=swa_metadata,
+                swa_k_cache=self.swa_cache_layer.kv_cache,
                 swa_only=swa_only,
                 topk_indices=topk_indices,
                 topk_lens=topk_lens,
                 swa_indices=swa_indices,
                 swa_lens=swa_lens,
+                attn_sink=self.attn_sink,
+                scale=self.scale,
+                head_dim=self.head_dim,
+                nope_head_dim=self.nope_head_dim,
+                rope_head_dim=self.rope_head_dim,
                 output=output,
             )
             return
@@ -1160,13 +1023,16 @@ class DeepseekV4MLAAttention(nn.Module, AttentionLayerBase):
             )
 
             if current_platform.is_rocm():
-                output_chunk = self._ref_sparse_attn_prefill(
+                rocm_sparse_attn_prefill(
                     q=q[query_start:query_end],
                     kv=kv.view(-1, 1, q.shape[-1]),
                     indices=combined_indices.unsqueeze(1),
                     topk_length=combined_lens,
+                    scale=self.scale,
+                    head_dim=self.head_dim,
+                    attn_sink=self.attn_sink,
+                    output=output[query_start:query_end],
                 )
-                output[query_start:query_end].copy_(output_chunk.to(output.dtype))
             else:
                 output_chunk, _, _ = flash_mla_sparse_fwd(
                     q=q[query_start:query_end],
@@ -1177,225 +1043,6 @@ class DeepseekV4MLAAttention(nn.Module, AttentionLayerBase):
                     topk_length=combined_lens,
                     out=output[query_start:query_end],
                 )
-
-    def _decode_e8m0_scales(self, scale: torch.Tensor) -> torch.Tensor:
-        from vllm.model_executor.layers.quantization.utils.fp8_utils import (
-            _upcast_e8m0_to_fp32,
-        )
-
-        return _upcast_e8m0_to_fp32(scale.contiguous())
-
-    def _dequantize_cache_rows(self, rows: torch.Tensor) -> torch.Tensor:
-        rows = rows.reshape(-1, rows.shape[-1])
-        fp8_dtype = current_platform.fp8_dtype()
-        fp8_dim = self.nope_head_dim
-        rope_bytes = self.rope_head_dim * 2
-        scale_dim = fp8_dim // 64
-
-        fp8_vals = rows[:, :fp8_dim].contiguous().view(fp8_dtype)
-        rope_vals = (
-            rows[:, fp8_dim : fp8_dim + rope_bytes].contiguous().view(torch.bfloat16)
-        )
-        scale_bytes = rows[
-            :, fp8_dim + rope_bytes : fp8_dim + rope_bytes + scale_dim
-        ].contiguous()
-        scales = self._decode_e8m0_scales(scale_bytes)
-        scales = torch.repeat_interleave(scales, 64, dim=-1)
-        nope = fp8_vals.to(torch.float32) * scales
-        return torch.cat([nope, rope_vals.to(torch.float32)], dim=-1).to(torch.bfloat16)
-
-    def _gather_dequantized_cache_tokens(
-        self,
-        cache: torch.Tensor,
-        slot_ids: torch.Tensor,
-        block_size: int,
-    ) -> torch.Tensor:
-        if slot_ids.numel() == 0:
-            return torch.empty(
-                (0, self.head_dim), dtype=torch.bfloat16, device=cache.device
-            )
-        slot_ids = slot_ids.to(torch.int64)
-        rows = cache[slot_ids // block_size, slot_ids % block_size]
-        return self._dequantize_cache_rows(rows).reshape(-1, self.head_dim)
-
-    def _forward_decode_fallback(
-        self,
-        q: torch.Tensor,
-        kv_cache: torch.Tensor | None,
-        swa_metadata: "DeepseekSparseSWAMetadata",
-        swa_only: bool,
-        topk_indices: torch.Tensor | None,
-        topk_lens: torch.Tensor | None,
-        swa_indices: torch.Tensor,
-        swa_lens: torch.Tensor,
-        output: torch.Tensor,
-    ) -> None:
-        blocked_swa = self._dequantize_blocked_k_cache(self.swa_cache_layer.kv_cache)
-        blocked_extra = None if swa_only else self._dequantize_blocked_k_cache(kv_cache)
-        attn_out = self._ref_sparse_attn_decode(
-            q=q.unsqueeze(1),
-            blocked_k=blocked_swa,
-            indices_in_kvcache=swa_indices.unsqueeze(1),
-            topk_length=swa_lens,
-            attn_sink=self.attn_sink[: q.shape[1]],
-            extra_blocked_k=blocked_extra,
-            extra_indices_in_kvcache=topk_indices,
-            extra_topk_length=topk_lens,
-        )
-        output.copy_(attn_out.to(output.dtype))
-
-    def _dequantize_blocked_k_cache(self, quant_k_cache: torch.Tensor) -> torch.Tensor:
-        fp8_dtype = current_platform.fp8_dtype()
-        d = self.head_dim
-        d_nope = self.nope_head_dim
-        d_rope = self.rope_head_dim
-        tile_size = 64
-        num_tiles = d_nope // tile_size
-
-        num_blocks, block_size, _ = quant_k_cache.shape
-        quant_k_cache = quant_k_cache.view(num_blocks, -1)
-        input_nope_rope = quant_k_cache[:, : block_size * (d_nope + 2 * d_rope)].view(
-            num_blocks, block_size, d_nope + 2 * d_rope
-        )
-        input_nope = input_nope_rope[:, :, :d_nope].view(fp8_dtype)
-        input_rope = input_nope_rope[:, :, d_nope:].view(torch.bfloat16)
-        input_scale = (
-            quant_k_cache[:, block_size * (d_nope + 2 * d_rope) :]
-            .view(num_blocks, block_size, 8)[:, :, :num_tiles]
-            .view(torch.float8_e8m0fnu)
-        )
-
-        result = torch.empty(
-            (num_blocks, block_size, 1, d),
-            dtype=torch.bfloat16,
-            device=quant_k_cache.device,
-        )
-        result[..., d_nope:] = input_rope.unsqueeze(2)
-        for tile_idx in range(num_tiles):
-            cur_nope = input_nope[
-                ..., tile_idx * tile_size : (tile_idx + 1) * tile_size
-            ].to(torch.bfloat16)
-            cur_scales = input_scale[:, :, tile_idx].to(torch.bfloat16).unsqueeze(-1)
-            result[..., tile_idx * tile_size : (tile_idx + 1) * tile_size] = (
-                cur_nope * cur_scales
-            ).unsqueeze(2)
-        return result
-
-    def _ref_sparse_attn_prefill(
-        self,
-        q: torch.Tensor,
-        kv: torch.Tensor,
-        indices: torch.Tensor,
-        topk_length: torch.Tensor | None,
-    ) -> torch.Tensor:
-        indices = indices.clone().squeeze(1)
-        s_q, h_q, d_qk = q.shape
-        topk = indices.shape[-1]
-        s_kv = kv.shape[0]
-        if topk_length is not None:
-            mask = torch.arange(topk, device=indices.device).unsqueeze(
-                0
-            ) >= topk_length.unsqueeze(1)
-            indices[mask] = -1
-        invalid_mask = (indices < 0) | (indices >= s_kv)
-        indices[invalid_mask] = 0
-
-        qf = q.float()
-        gathered_kv = (
-            kv.index_select(0, indices.flatten()).reshape(s_q, topk, d_qk).float()
-        )
-        scores = qf @ gathered_kv.transpose(1, 2)
-        scores *= self.scale
-        scores[invalid_mask.unsqueeze(1).expand_as(scores)] = float("-inf")
-
-        orig_lse = torch.logsumexp(scores, dim=-1)
-        lse_for_o = orig_lse
-        if self.attn_sink is not None:
-            lse_for_o = torch.logsumexp(
-                torch.stack(
-                    [orig_lse, self.attn_sink[:h_q].view(1, h_q).expand_as(orig_lse)],
-                    dim=0,
-                ),
-                dim=0,
-            )
-        lse_for_o = lse_for_o.clone()
-        lse_for_o[lse_for_o == float("-inf")] = float("+inf")
-        probs = torch.exp(scores - lse_for_o.unsqueeze(-1))
-        out = probs @ gathered_kv[..., : self.head_dim]
-        lonely_q_mask = orig_lse == float("-inf")
-        out[lonely_q_mask.unsqueeze(-1).expand_as(out)] = 0.0
-        return out.to(torch.bfloat16)
-
-    def _ref_sparse_attn_decode(
-        self,
-        q: torch.Tensor,
-        blocked_k: torch.Tensor,
-        indices_in_kvcache: torch.Tensor,
-        topk_length: torch.Tensor | None,
-        attn_sink: torch.Tensor | None,
-        extra_blocked_k: torch.Tensor | None = None,
-        extra_indices_in_kvcache: torch.Tensor | None = None,
-        extra_topk_length: torch.Tensor | None = None,
-    ) -> torch.Tensor:
-        b, s_q, h_q, d_qk = q.shape
-        d_v = self.head_dim
-
-        def process_scope(
-            cur_blocked_k: torch.Tensor,
-            cur_indices: torch.Tensor,
-            cur_topk_length: torch.Tensor | None,
-        ) -> tuple[torch.Tensor, torch.Tensor]:
-            cur_indices = cur_indices.reshape(b, s_q, -1)
-            topk = cur_indices.size(-1)
-            fixed_indices = torch.clamp_min(cur_indices, 0)
-            gathered_kv = (
-                cur_blocked_k.view(-1, d_qk)
-                .index_select(0, fixed_indices.view(-1))
-                .view(b, s_q, topk, d_qk)
-            )
-            invalid_mask = cur_indices == -1
-            if cur_topk_length is not None:
-                cur_topk_length = cur_topk_length.reshape(b)
-                invalid_mask |= torch.arange(0, topk, device=invalid_mask.device).view(
-                    1, 1, topk
-                ) >= cur_topk_length.view(b, 1, 1)
-            return gathered_kv, invalid_mask
-
-        gathered_kv, invalid_mask = process_scope(
-            blocked_k, indices_in_kvcache, topk_length
-        )
-        if extra_blocked_k is not None:
-            assert extra_indices_in_kvcache is not None
-            gathered_kv1, invalid_mask1 = process_scope(
-                extra_blocked_k, extra_indices_in_kvcache, extra_topk_length
-            )
-            gathered_kv = torch.cat([gathered_kv, gathered_kv1], dim=2)
-            invalid_mask = torch.cat([invalid_mask, invalid_mask1], dim=2)
-
-        gathered_kv = gathered_kv.view(b * s_q, -1, d_qk).float()
-        gathered_kv[gathered_kv != gathered_kv] = 0.0
-        qf = q.float().view(b * s_q, h_q, d_qk)
-        attn_weight = qf @ gathered_kv.transpose(-1, -2)
-        attn_weight *= self.scale
-        attn_weight[
-            invalid_mask.view(b * s_q, 1, -1).expand(
-                b * s_q, h_q, invalid_mask.size(-1)
-            )
-        ] = float("-inf")
-        lse = attn_weight.logsumexp(dim=-1)
-        attn_weight = torch.exp(attn_weight - lse.unsqueeze(-1))
-        output = attn_weight @ gathered_kv[..., :d_v]
-        output = output.view(b, s_q, h_q, d_v)
-        lse = lse.view(b, s_q, h_q)
-
-        if attn_sink is not None:
-            output *= (
-                1.0 / (1.0 + torch.exp(attn_sink.view(1, 1, h_q) - lse))
-            ).unsqueeze(-1)
-
-        lonely_q_mask = lse == float("-inf")
-        output[lonely_q_mask.unsqueeze(-1).expand_as(output)] = 0.0
-        return output.squeeze(1).to(torch.bfloat16)
 
 
 class DeepseekV4IndexerCache(torch.nn.Module, AttentionLayerBase):
